pyproject.toml
[project]
name = "hackathon-umass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.60.0",
    "dotenv>=0.9.9",
    "elevenlabs>=2.22.0",
    "google-genai>=1.49.0",
    "gradio>=3.36.1",
    "mss>=10.1.0",
    "openai>=2.7.1",
    "pillow>=10.0.0,<12.0.0",
    "pydantic>=2.0.0",
    "pync>=2.0.3",
    "pynput>=1.8.1",
    "setuptools>=80.9.0",
]
